server/requirements.txt
Flask==3.0.3
httpx==0.28.1
openai==1.65.4
pydantic==2.10.6
python-dotenv==1.0.1
